Refactored REST API packages in prep for new runtime API
