+++ bxbot-rest-api/src/main/java/com/gazbert/bxbot/rest/api/config/EngineConfigController.java
@@ -21,10 +21,10 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.gazbert.bxbot.rest.api.emailalerts;
+package com.gazbert.bxbot.rest.api.config;
 
-import com.gazbert.bxbot.domain.emailalerts.EmailAlertsConfig;
-import com.gazbert.bxbot.services.EmailAlertsConfigService;
+import com.gazbert.bxbot.domain.engine.EngineConfig;
+import com.gazbert.bxbot.services.EngineConfigService;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -39,58 +39,57 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
+ * Controller for directing Engine config requests.
  * <p>
- * Controller for directing Email Alerts config requests.
+ * Engine config can only be fetched and updated - it cannot be deleted or created.
  * <p>
- * Email Alerts config can only be fetched and updated - it cannot be deleted or created.
- * <p>
- * There is only 1 Email Alerter per bot.
+ * There is only 1 Trading Engine per bot.
  *
  * @author gazbert
  * @since 1.0
  */
 @RestController
 @RequestMapping("/api/config")
-public class EmailAlertsConfigController {
+public class EngineConfigController {
 
     private static final Logger LOG = LogManager.getLogger();
-    private final EmailAlertsConfigService emailAlertsConfigService;
+    private final EngineConfigService engineConfigService;
 
     @Autowired
-    public EmailAlertsConfigController(EmailAlertsConfigService emailAlertsConfigService) {
-        Assert.notNull(emailAlertsConfigService, "emailAlertsConfigService dependency cannot be null!");
-        this.emailAlertsConfigService = emailAlertsConfigService;
+    public EngineConfigController(EngineConfigService engineConfigService) {
+        Assert.notNull(engineConfigService, "engineConfigService dependency cannot be null!");
+        this.engineConfigService = engineConfigService;
     }
 
     /**
-     * Returns the Email Alerts configuration for the bot.
+     * Returns the Engine configuration for the bot.
      *
-     * @return the Email Alerts configuration.
+     * @return the Engine configuration.
      */
-    @RequestMapping(value = "/email-alerts", method = RequestMethod.GET)
-    public EmailAlertsConfig getEmailAlerts(@AuthenticationPrincipal User user) {
+    @RequestMapping(value = "/engine", method = RequestMethod.GET)
+    public EngineConfig getEngine(@AuthenticationPrincipal User user) {
 
-        LOG.info("GET /email-alerts - getEmailAlerts() - caller: " + user.getUsername());
+        LOG.info("GET /engine - getEngine() - caller: " + user.getUsername());
 
-        final EmailAlertsConfig emailAlertsConfig = emailAlertsConfigService.getEmailAlertsConfig();
+        final EngineConfig engineConfig = engineConfigService.getEngineConfig();
 
-        LOG.info("Response: " + emailAlertsConfig);
-        return emailAlertsConfig;
+        LOG.info("Response: " + engineConfig);
+        return engineConfig;
     }
 
     /**
-     * Updates the Email Alerts configuration for the bot.
+     * Updates the Engine configuration for the bot.
      *
-     * @return 200 'OK' HTTP status code and Email Alerts config in response body if update successful,
+     * @return 200 'OK' HTTP status code and updated Engine config in the response body if update successful,
      *         some other HTTP status code otherwise.
      */
-    @RequestMapping(value = "/email-alerts", method = RequestMethod.PUT)
-    public ResponseEntity<?> updateEmailAlerts(@AuthenticationPrincipal User user, @RequestBody EmailAlertsConfig config) {
+    @RequestMapping(value = "/engine", method = RequestMethod.PUT)
+    public ResponseEntity<?> updateEngine(@AuthenticationPrincipal User user, @RequestBody EngineConfig config) {
 
-        LOG.info("PUT /email-alerts - updateEmailAlerts() - caller: " + user.getUsername());
+        LOG.info("PUT /engine - updateEngine() - caller: " + user.getUsername());
         LOG.info("Request: " + config);
 
-        final EmailAlertsConfig updatedConfig = emailAlertsConfigService.updateEmailAlertsConfig(config);
+        final EngineConfig updatedConfig = engineConfigService.updateEngineConfig(config);
         return new ResponseEntity<>(updatedConfig, HttpStatus.OK);
     }
 }
