+++ bxbot-rest-api/src/main/java/com/gazbert/bxbot/rest/api/config/ExchangeConfigController.java
@@ -21,7 +21,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.gazbert.bxbot.rest.api.exchange;
+package com.gazbert.bxbot.rest.api.config;
 
 import com.gazbert.bxbot.domain.exchange.ExchangeConfig;
 import com.gazbert.bxbot.services.ExchangeConfigService;
