+++ bxbot-rest-api/src/main/java/com/gazbert/bxbot/rest/api/config/MarketConfigController.java
@@ -21,7 +21,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.gazbert.bxbot.rest.api.markets;
+package com.gazbert.bxbot.rest.api.config;
 
 import com.gazbert.bxbot.domain.market.MarketConfig;
 import com.gazbert.bxbot.services.MarketConfigService;
