+++ bxbot-rest-api/src/main/java/com/gazbert/bxbot/rest/api/config/StrategyConfigController.java
@@ -21,7 +21,7 @@
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.gazbert.bxbot.rest.api.strategies;
+package com.gazbert.bxbot.rest.api.config;
 
 import com.gazbert.bxbot.domain.strategy.StrategyConfig;
 import com.gazbert.bxbot.services.StrategyConfigService;
