+++ bxbot-rest-api/src/main/java/com/gazbert/bxbot/rest/api/config/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * <h2>BX-bot Configuration REST API</h2>
+ * <p>
+ * This is the REST API for administering the bot's configuration.
+ *
+ * @author gazbert
+ * @since 1.0
+ */
+package com.gazbert.bxbot.rest.api.config;
