--- bxbot-rest-api/src/main/java/com/gazbert/bxbot/rest/api/package-info.java
@@ -1,9 +1,0 @@
-/**
- * <h2>BX-bot REST API</h2>
- * <p>
- * This is the REST API for administering the bot.
- *
- * @author gazbert
- * @since 1.0
- */
-package com.gazbert.bxbot.rest.api;
